Experimental refactoring of MongoCollection API. Now the Collection classes are:
MongoCollection (base class)
  MongoCollectionBsonDocument (used with BsonDocuments)
  MongoCollection<TDocument> (used with POCOs whose type is supplied at compile time)
  MongoCollectionTyped (used with POCOs whose type is supplied at runtime)

The existing GetCollection methods in MongoDatabase have been altered to return one of the three new subclasses (the parameters of the overload determine which one is returned).

The following methods have been spread around in the new derived classes:

  all Find methods
  all Geo search methods
  all Insert and InsertBatch methods
  all Save methods

There are now a lot fewer overloads because the methods have been distributed among the subclasses.

Also, methods that used to end in "As" have either been removed or renamed by stripping "As" from the end of the name.

There has also been a bit of work on how (and where) RunCommand is implemented.